Update patch set 1
Patch Set 1:

I think those errors aren't a result of the USB config but related to some of the VAB handling, especially the merge status error. There are presently some issues around merges and formatting of data that still need to be resolved. I think the USB config may have seemed like the problem but is unrelated.

Patch-set: 1